Colors still in progress

+++ fdf.h
@@ -52,10 +52,10 @@ typedef struct	s_map
 }				t_map;
 
 // FDF 6: Color
-int				set_color(t_map *map, int z);
+int				set_color(t_map *map, double z);
 
 // FDF 2: Drawing
-void			draw_pixel(t_map *map, int x, int y, int z);
+void			draw_pixel(t_map *map, int x, int y, double z);
 void			draw_line_v(t_pt *a, t_pt *b, t_map *map);
 void			draw_line_h(t_pt *a, t_pt *b, t_map *map);
 void			draw_line(t_pt *a, t_pt *b, t_map *map);
